requirements.txt
# Stock Data Explorer Dependencies

# Core application framework
streamlit>=1.30.0

# HTTP requests for API calls
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Testing frameworks
pytest>=7.4.0
hypothesis>=6.92.0

# Data visualization (matplotlib as backup for plotting)
matplotlib>=3.8.0
